reinstall only on deploy

+++ fabfile.py
@@ -12,8 +12,9 @@ ALL_CONFIGS = frozenset(['BOT_USERNAME', 'BOT_API_KEY'])
 
 
 @runs_once
-def install_xlib():
-    local('rm -rf xlib/*')
+def install_xlib(reinstall=True):
+    if reinstall:
+        local('rm -rf xlib/*')
     local('printf \'%s\n%s\n\' \'[install]\' \'prefix=\' > ~/.pydistutils.cfg')
     local('pip install --upgrade --no-deps --requirement requirements_xlib.txt -t xlib')
     local('rm ~/.pydistutils.cfg')
@@ -63,7 +64,7 @@ def prepare_env(project=None):
 def debug():
     """Starts up debug server"""
     install()
-    install_xlib()
+    install_xlib(reinstall=False)
     prepare_env()
     local('dev_appserver.py --host 0.0.0.0 --log_level debug app.yaml')
